backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
mysql-connector-python==8.3.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.9
pydantic==2.5.3 